+++ .gitignore
@@ -38,3 +38,4 @@ next-env.d.ts
 # wrangler files
 .wrangler
 .dev.vars
+wrangler.toml
