+++ components/_AppShell.module.css
@@ -1,0 +1,25 @@
+.icon {
+    width: 22px;
+    height: 22px;
+  }
+  
+  .dark {
+    @mixin dark {
+      display: none;
+    }
+  
+    @mixin light {
+      display: block;
+    }
+  }
+  
+  .light {
+    @mixin light {
+      display: none;
+    }
+  
+    @mixin dark {
+      display: block;
+    }
+  }
+
