+++ package.json
@@ -15,6 +15,7 @@
   "dependencies": {
     "@mantine/core": "^7.11.2",
     "@mantine/hooks": "^7.11.2",
+    "@tabler/icons-react": "^3.12.0",
     "next": "^14.2.5",
     "react": "^18",
     "react-dom": "^18"
@@ -25,8 +26,8 @@
     "@types/node": "^20",
     "@types/react": "^18",
     "@types/react-dom": "^18",
-    "postcss": "^8.4.39",
-    "postcss-preset-mantine": "^1.16.0",
+    "postcss": "^8.4.41",
+    "postcss-preset-mantine": "^1.17.0",
     "postcss-simple-vars": "^7.0.1",
     "typescript": "^5",
     "vercel": "^32.3.0",
